Migrated code from scripts.js to dom.js, updated styling for friends dropdown
