+++ src/dom.js
@@ -1,3 +1,20 @@
+// -------------------SCSS Imports----------------------
+import './css/base.scss';
+import './css/styles.scss';
+
+// -------------------NPM Package Imports----------------------
+import dayjs from 'dayjs';
+import '@fortawesome/fontawesome-free/js/fontawesome'
+import '@fortawesome/fontawesome-free/js/solid'
+import '@fortawesome/fontawesome-free/js/regular'
+import '@fortawesome/fontawesome-free/js/brands'
+
+// -------------------Class Imports----------------------
+import Activity from './Activity';
+import Hydration from './Hydration';
+import Sleep from './Sleep';
+
+// -------------------Chart Imports----------------------
 import { renderWaterConsumed } from './charts/hydration-charts/water-consumed-chart.js'
 import { renderWaterOverWeek } from './charts/hydration-charts/water-over-week.js'
 import { renderStepsMiles } from './charts/activity-charts/steps-miles-chart.js'
@@ -9,153 +26,312 @@ import { renderUserAnalyticsVsAll } from './charts/activity-charts/activity-anal
 import { renderSleepOverWeek } from './charts/sleep-charts/weekly-sleep-chart.js'
 import { renderWeeklyActivity } from './charts/activity-charts/weekly-activity-analytics-chart.js'
 
+// -----------------------GLOBAL VARIABLES--------------------
+let picker, user;
+// -----------------------QUERY SELECTORS---------------------------
+const userCard = document.getElementById('userinfo');
+const friends = document.getElementById('dropDownContent');
+const friendsBtn = document.getElementById('friendsDropDown');
+// PAGE LOAD
 
-const friendsDiv = document.getElementById('dropDownContent'); // CANT GET DIV TO WORK WITHOUT THIS HERE(used to be called 'const friends' only)
-
-const renderCharts = (currentUser, currentDate, userRepo, firstDate) => {
-  renderWaterConsumed(currentUser, currentDate.date)
-  renderWaterOverWeek(currentUser, currentDate.date)
-  renderStepsMiles(currentUser, currentDate.date)
-  renderAllTimeSleep(currentUser, currentDate.date)
-  renderSleepQuality(currentUser, currentDate.date)
+export const renderPage = (currentUser, currentDate, userRepo, firstDate) => {
+  renderWaterConsumed(currentUser, currentDate)
+  renderWaterOverWeek(currentUser, currentDate)
+  renderStepsMiles(currentUser, currentDate)
+  renderAllTimeSleep(currentUser, currentDate)
+  renderSleepQuality(currentUser, currentDate)
   renderUserCard(currentUser);
   renderUserStepGoalVsAverage(currentUser, userRepo);
-  renderLastMinActive(currentUser, currentDate.date);
-  renderUserAnalyticsVsAll(currentUser, currentDate.date, userRepo)
-  renderSleepOverWeek(currentUser, currentDate.date)
-  renderWeeklyActivity(currentUser, currentDate.date)
-  renderDatePicker(currentDate.date, firstDate.date)
+  renderLastMinActive(currentUser, currentDate);
+  renderUserAnalyticsVsAll(currentUser, currentDate, userRepo)
+  renderSleepOverWeek(currentUser, currentDate)
+  renderWeeklyActivity(currentUser, currentDate)
+  renderDatePicker(currentDate, firstDate, currentUser, userRepo)
+  renderFriends(currentUser, userRepo);
 }
-
-const renderFriends = (currentUser) => {
-  friendsDiv.innerHTML = 'test';
-  let userFriends = currentUser.friends.map(friend => {
-      return userRepo.users.filter(user => {
-          if (user.id === friend) {
-            return user
-          }
-      })
-  }).flat();
-  friendsDiv.innerHTML +=
-    `<p class='friend-details'>${userFriends.map(friend => {return `${friend.name} | ${friend.dailyStepGoal} steps` + "<br>"}).join('')}</p>`
-};
 
 const toggleFriends = () => {
-    friendsDiv.classList.toggle('hidden')
-  }
-
-const renderDatePicker = (currentDate, firstDate) => {
-
-  let current = new Date(currentDate)
-  let min = new Date(firstDate)
-  let picker = new Pikaday({
-   field: document.getElementById('datePicker'),
-   defaultDate: current,
-   minDate: min,
-   maxDate: current,
- })
+  friends.classList.toggle('hidden')
 }
 
-const renderUserCard = (currentUser) => {
-  const userCard = document.getElementById('userinfo'); //This gonna have to stay here?
-  userCard.innerHTML =
-  `   <article id='user' width="300" height="300">
-          <div class='user-greeting'>
-            <h1>Welcome back, ${currentUser.name.split(' ')[0]}!</h1>
-          </div>
-          <section class='user-details' id='userDetails'>
-              <div class='user-address' id='userAddress'>
-                <p><strong> ADDRESS: </strong>${currentUser.address}</p>
-              </div>
-              <div class='user-email' id='userEmail'>
-                <p><strong> EMAIL: </strong>${currentUser.email}</p>
-              </div>
-              <div class='user-step-goal' id='userStepGoal'>
-                <p><strong> DAILY STEP GOAL: </strong>${currentUser.dailyStepGoal}</p>
-              </div>
-          </section>
-      </article> `
-};
+export const renderFriends = (currentUser, userRepo) => {
+  friendsBtn.addEventListener('click', toggleFriends)
+  friends.innerHTML = '';
+  let userFriends = currentUser.friends.map(friend => {
+    return userRepo.users.filter(user => {
+      if (user.id === friend) {
+        return user
+      }
+    })
+  }).flat();
+    friends.innerHTML +=
+    `<p class='friend-details'>${userFriends.map(friend => {return `${friend.name} | ${friend.dailyStepGoal} steps` + "<br>"}).join('')}</p>`
+  };
 
+  const renderUserCard = (currentUser) => {
+    userCard.innerHTML = `
+      <article id='user' width="300" height="300">
+        <div class='user-greeting'>
+          <h1>Welcome back, ${currentUser.name.split(' ')[0]}!</h1>
+        </div>
+        <section class='user-details' id='userDetails'>
+            <div class='user-address' id='userAddress'>
+              <p><strong> ADDRESS: </strong>${currentUser.address}</p>
+            </div>
+            <div class='user-email' id='userEmail'>
+              <p><strong> EMAIL: </strong>${currentUser.email}</p>
+            </div>
+            <div class='user-step-goal' id='userStepGoal'>
+              <p><strong> DAILY STEP GOAL: </strong>${currentUser.dailyStepGoal}</p>
+            </div>
+        </section>
+      </article>
+    `
+  };
+
+  export const renderDatePicker = (currentDate, firstDate, currentUser, userRepo) => {
+    const testInput = (value) => {
+      alert(`THE VALUE IS ${value}`)
+    }
+    let current = new Date(currentDate)
+    let min = new Date(firstDate)
+    picker = new Pikaday({
+     field: document.getElementById('datePicker'),
+     defaultDate: current,
+     minDate: min,
+     maxDate: current,
+   })
+    document.getElementById('datePicker').addEventListener('change', (event) => {
+      let value = event.target.value
+      let day = value.split(' ').slice(1, 4).join(' ')
+      let formatted = dayjs(day).format("YYYY/MM/DD")
+      resetCharts();
+      renderCanvases()
+      updateCharts(currentUser, formatted, userRepo)
+   });
+  }
+
+const updateCharts = (currentUser, formatted, userRepo) => {
+  renderWaterConsumed(currentUser, formatted)
+  renderWaterOverWeek(currentUser, formatted)
+  renderStepsMiles(currentUser, formatted)
+  renderAllTimeSleep(currentUser, formatted)
+  renderSleepQuality(currentUser, formatted)
+  renderUserStepGoalVsAverage(currentUser, userRepo);
+  renderLastMinActive(currentUser, formatted);
+  renderUserAnalyticsVsAll(currentUser, formatted, userRepo)
+  renderSleepOverWeek(currentUser, formatted)
+  renderWeeklyActivity(currentUser, formatted)
+}
+
+  export const resetCharts = () => {
+    let allChartContainers = document.querySelectorAll('.chart-ctr')
+    allChartContainers.forEach(container => {
+      container.innerHTML = ""
+    })
+  }
+
+  export const renderCanvases = () => {
+    document.getElementById('waterconsumed').innerHTML += `
+    <canvas id="water-chart" width="300" height="300"></canvas>
+    `
+    document.getElementById('wateroverweek').innerHTML += `
+    <canvas id="waterWeekChart" width="300" height="300"></canvas>
+    `
+    document.getElementById('latestsleepdata').innerHTML += `
+    <canvas id="sleep-chart" width="300" height="300"></canvas>
+    `
+    document.getElementById('latestsleepoverweek').innerHTML += `
+    <canvas id="sleepOverWeekChart" width="300" height="300"></canvas>
+    `
+    document.getElementById('alltimesleepavg').innerHTML += `
+    <canvas id="allTimeSleep-chart" ></canvas>
+    `
+    document.getElementById('lateststepsmiles').innerHTML += `
+    <canvas id="stepsMilesChart" width="300" height="300"></canvas>
+    `
+    document.getElementById('usergoalvsavg').innerHTML += `
+    <canvas id="useravgchart"></canvas>
+    `
+    document.getElementById('activityvsuserbase').innerHTML += `
+    <canvas id="userAnalyticsVsUserBase" width="626" height="312"></canvas>
+    `
+    document.getElementById('weeklyactivitydata').innerHTML += `
+    <div id="min-stairs-container">
+      <canvas id="minStairsAvgChart" width="224" height="300"></canvas>
+    </div>
+    <div id="steps-container">
+      <canvas id="stepsAvgChart"  width="224" height="300"></canvas>
+    </div>  `
+  }
+
+  function postData(dataType, body) {
+    const root = 'http://localhost:3001/api/v1/'
+    fetch(root + dataType, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(body),
+    })
+      .then(response => console.log(response.status))
+      .catch(err => console.error(err))
+  }
+
+  function createSleepBody() {
+    let userSleepDate = document.getElementById('sleep-input-date').value
+    let userHoursSlept = parseFloat(document.getElementById('input-hours-slept').value)
+    let userSleepQuality = parseFloat(document.getElementById('input-sleep-quality').value)
+    let sleep = {userID: currentUser.id, date: userSleepDate, hoursSlept: userHoursSlept, sleepQuality: userSleepQuality}
+    return new Sleep(sleep)
+  }
+
+  function createHydrationBody() {
+    let userHydrationDate = document.getElementById('hydration-input-date').value
+    let userOuncesConsumed = parseFloat(document.getElementById('input-ounces-number').value)
+    let hydration = {userID: currentUser.id, date: userHydrationDate, numOunces: userOuncesConsumed}
+    return new Hydration(hydration)
+  }
+
+  function createActivityBody() {
+    let userActivityDate = document.getElementById('activity-input-date').value
+    let userNumberOfSteps = parseFloat(document.getElementById('input-step-number').value)
+    let userMinutesActive = parseFloat(document.getElementById('input-minutes-active').value)
+    let userStairsClimbed = parseFloat(document.getElementById('input-stairs-climbed').value)
+    let activity = {userID: currentUser.id, date: userActivityDate, numSteps: userNumberOfSteps, minutesActive: userMinutesActive, flightsOfStairs: userStairsClimbed}
+    return new Activity(activity)
+  }
+
+  function formSubmitClickHandler(event) {
+    if (event.target.id === 'submit-sleep') {
+      let sleepBody = createSleepBody();
+      postData('sleep', sleepBody)
+      closeModal();
+    }
+    if (event.target.id === 'submit-hydration') {
+      let hydrationBody = createHydrationBody();
+      postData('hydration', hydrationBody)
+      closeModal();
+    }
+    if (event.target.id === 'submit-activity') {
+      let activityBody = createActivityBody();
+      postData('activity', activityBody)
+      closeModal();
+    }
+  }
+
+  const addEntryBtn = document.getElementById('addEntryBtn');
+  const userInputModal = document.getElementById('userInputModal');
 
   const getUserInput = () => {
-  userInputModal.innerHTML = '';
-  userInputModal.innerHTML +=
+    userInputModal.innerHTML = '';
+    userInputModal.innerHTML +=
     `<article class='user-input-content'>
-        <button class='close-modal' id='close'>
+        <button class='close-modal' id='close' aria-label="close-modal">
           <i class="far fa-times-circle"></i>
         </button>
-        <h1 class='user-input-header'>Add New Fitness Data</h1>
-          <form class='user-input-sleep' id='userInputSleep'>
-            <h2>Add New Sleep Data</h2>
-            <label for="sleep-user-date">Date</label>
-            <input type="text" name="sleep-user-input-date" id="sleep-input-date" placeholder="yyyy/mm/dd">
-            <label for="user-hours-slept">Hours Slept</label>
-            <input type="number" name="user-hours-slept" id="input-hours-slept" min="0" max="40">
-            <label for="user-sleep-quality">Sleep Quality</label>
-            <input type="number" name="user-sleep-quality" id="input-sleep-quality" min="1" max="5" placeholder="Enter a number from 1-5" step=".1">
-            <input value="Submit" class="submit-info" id="submit-sleep" type="submit">
-          </form>
-          <form class='user-input-activity' id='userInputActivity'>
-            <h2>Add New Activity Data</h2>
-            <label for="activity-user-date">Date</label>
-            <input type="text" name="activity-user-date" id="activity-input-date" placeholder="yyyy/mm/dd">
-            <label for="user-step-number">Number of Steps</label>
-            <input type="number" name="user-step-number" id="user-step-number" mix="0">
-            <label for="user-minutes-active">Active Minutes</label>
-            <input type="number" name="user-minutes-active" id="user-minutes-active" min="0">
-            <label for="user-stairs-climbed">Flight of Stairs Climbed</label>
-            <input type="number" name="user-stairs-climbed" id="user-stairs-climbed" min="0">
-            <button class="submit-info" id="submit-activity">Submit</button>
-          </form>
-          <form class='user-input-hydration' id='userInputHydration'>
-            <h2>Add New Hydration Data</h2>
-            <label for="hydration-user-date">Date</label>
-            <input type="text" name="hydration-user-date" id="hydration-input-date" placeholder="yyyy/mm/dd">
-            <label for="user-ounces-number">Ounces of Water Drank</label>
-            <input type="number" name="user-ounces-number" id="input-ounces-number" min="0">
-            <button class="submit-info" id="submit-hydration">Submit</button>
-          </form>
-    </article>`
-  openModal()
-}
+        <section class='user-input-form' id='userInputForm'>
+          <h1 class='user-input-header'>What Would You Like To Add?</h1>
+          <div class='data-buttons' id='dataButtons'>
+            <button id='addSleep'>Sleep</button>
+            <button id='addWater'>Water</button>
+            <button id='addActivity'>Activity</button>
+          </div>
+        </section>
+     </article>`
+   openModal()
+  }
 
-const openModal = () => {
-  userInputModal.style.display = 'flex'
-}
+  addEntryBtn.addEventListener('click', () => {
+    getUserInput()
+  });
 
-const closeModal = () => {
-  console.log('closeModal function in domljs');
-  userInputModal.style.display = 'none';
-}
+  userInputModal.addEventListener('click', (event) => {
+    modalClickHandler(event)
+  });
 
-const modalClickHandler = (event) => {
-  event.preventDefault();
-  if(event.target.id === 'close') {
-    closeModal();
-  } else if (event.target.id === 'submit-activity' || event.target.id === 'submit-sleep' 
-    || event.target.id === 'submit-hydration') {
+  function modalClickHandler(event){
+    event.preventDefault();
+    if(event.target.id === 'close') {
+      console.log('click')
+      closeModal();
+    } else if (event.target.id === 'addSleep') {
+      renderSleepInputField()
+    } else if (event.target.id === 'addWater') {
+      renderWaterInputField()
+    } else if (event.target.id === 'addActivity') {
+      renderActivityInputField()
+    } else if (event.target.id === 'submit-activity' || event.target.id === 'submit-sleep' || event.target.id === 'submit-hydration') {
       formSubmitClickHandler(event);
       closeModal();
+    } else if (event.target.id === 'returnToSelection') {
+      getUserInput();
+    }
   }
-}
 
-const formSubmitClickHandler =(event) => {
-  if (event.target.id === 'submit-sleep') {
-    let sleepBody = createSleepBody();
-    postData('sleep', sleepBody)
-    closeModal();
+  function openModal() {
+    userInputModal.style.display = 'flex'
   }
-  if (event.target.id === 'submit-hydration') {
-    let hydrationBody = createHydrationBody();
-    postData('hydration', hydrationBody)
-    closeModal();
-  }
-  if (event.target.id === 'submit-activity') {
-    let activityBody = createActivityBody();
-    postData('activity', activityBody)
-    closeModal();
-  }
-}
 
-export { getUserInput, openModal, closeModal, modalClickHandler, renderCharts, renderFriends, toggleFriends }
+  function closeModal() {
+    userInputModal.style.display = 'none';
+  }
+
+  // let userInputForm = document.getElementById('userInputForm');
+
+  function renderWaterInputField() {
+    let userInputForm = document.getElementById('userInputForm')
+    userInputForm.innerHTML = ''
+    userInputForm.innerHTML +=
+    `<section class='user-input-form' id='userInputForm'>
+      <form class='user-input-hydration' id='userInputHydration'>
+        <h2>Add Water Intake</h2>
+        <label for="hydration-user-date">Date</label>
+        <input type="text" name="hydration-user-date" id="hydration-input-date" placeholder="yyyy/mm/dd" required>
+        <label for="user-ounces-number">Ounces of Water Drank</label>
+        <input type="number" name="user-ounces-number" id="input-ounces-number" min="0" required>
+        <button class="submit-info" id="submit-hydration">Submit</button>
+      </form>
+     </section>
+     <button class='return-to-selection' id='returnToSelection'>Go Back</button>`
+  }
+
+  function renderActivityInputField() {
+    let userInputForm = document.getElementById('userInputForm')
+    userInputForm.innerHTML = ''
+    userInputForm.innerHTML +=
+    `<section class='user-input-form' id='userInputForm'>
+      <form class='user-input-activity' id='userInputActivity'>
+        <h2>Add New Activity Data</h2>
+        <label for="activity-user-date">Date</label>
+        <input type="text" name="activity-user-date" id="activity-input-date" placeholder="yyyy/mm/dd">
+        <label for="user-step-number">Number of Steps</label>
+        <input type="number" name="user-step-number" id="user-step-number" mix="0">
+        <label for="user-minutes-active">Active Minutes</label>
+        <input type="number" name="user-minutes-active" id="user-minutes-active" min="0">
+        <label for="user-stairs-climbed">Flight of Stairs Climbed</label>
+        <input type="number" name="user-stairs-climbed" id="user-stairs-climbed" min="0">
+        <button class="submit-info" id="submit-activity">Submit</button>
+      </form>
+     </section>
+     <button class='return-to-selection' id='returnToSelection'>Go Back</button>`
+  }
+
+  function renderSleepInputField() {
+    let userInputForm = document.getElementById('userInputForm')
+    userInputForm.innerHTML = ''
+    userInputForm.innerHTML +=
+    `<section class='user-input-form' id='userInputForm'>
+      <form class='user-input-sleep' id='userInputSleep'>
+        <h2>Add New Sleep Data</h2>
+        <label for="sleep-user-date">Date</label>
+        <input type="text" name="sleep-user-input-date" id="sleep-input-date" placeholder="yyyy/mm/dd">
+        <label for="user-hours-slept">Hours Slept</label>
+        <input type="number" name="user-hours-slept" id="input-hours-slept" min="0" max="40">
+        <label for="user-sleep-quality">Sleep Quality</label>
+        <input type="number" name="user-sleep-quality" id="input-sleep-quality" min="1" max="5" placeholder="Enter a number from 1-5" step=".1">
+        <input value="Submit" class="submit-info" id="submit-sleep" type="submit">
+      </form>
+     </section>
+     <button class='return-to-selection' id='returnToSelection'>Go Back</button>`
+  }
